Add new transfomer and adjust test for other
To fit new requirements a new transformer was introduced and tests for an existing one
where improved.
The new transformer and the possibility to apply multiple transfomers was required as
we wanted to keep the size and identity of the validation set the same over all datasets
Do do this we need to shrink x_train, y_train to be able to partition it correctly.
As this triggered the assertions in the transfomer decorator we introduced the new
possibility to apply multiple transformers. take_balanced is therefore introduced
to modify xy_train so that it can be processed further