implement coordinator's state machine scaffold
The idea is to implement the core protocol without doing any explicit
IO, such that it becomes easily testable. In particular, timeouts are
injected from the outside.

That is maybe a slightly over-engineered solution for now, but, if we
are going to implement fancy error-tolerance, such reified states and
IO should become handy!

Note that the protocol is pretty simplistic at the moment:

* it requires all participants to submit their results (there's no
  selection)
* Participants are assumed not to fail
* Timed-out round just tears everything down
* No protection is made against late messages leaking between the
  rounds (it's not needed now, because we wait for all participants
  anyway)