XP-436 Reinstate FINISHED heartbeat from Coordinator (#219)
* XP-208 model sum aggregator as a mock for testing

* XP-208 add test integrating P and C

* XP-208 add identity controller as a mock for test

* XP-208 small typo in id-controller docstring

* XP-208 add coordinator test fixture with mocked components

* XP-208 test for top-level function start_participant

* XP-436 fix bug: coordinator not advertising FINISHED state

* XP-436 revert logging level

* XP-308: store aggregated weights in S3 buckets (#215)

* XP-308: store aggregated weights in S3 buckets

Refs:

https://xainag.atlassian.net/browse/XP-40
https://xainag.atlassian.net/browse/XP-308
https://xainag.atlassian.net/browse/XP-309
https://xainag.atlassian.net/browse/XP-311

Summary:

This commits adds a new module `xain-fl.coordinator.store`, which
contains a `Store()` class that can read from and write to S3
buckets. At the end of each round, the aggregated weights are uploaded
to an S3 bucket. Note that as a first step, the aggregated weights are
still sent to the participants via gRPC.

The versioning is straightforward: the key for storing each set of
weights is the round number corresponding to the weights.

Implementation notes:

We're trying to not have IO related code in the Coordinator class, so
the Store instance is held by the CoordinatorGrpc instead of
Coordinator. However, since the Coordinator is currently generating
the grpc messages, which are inherently coupled to the IO logic. This
means that if at some point the messages need to contain information
related to storage, we'll end up with storage logic in the
Coordinator, and things might become pretty messy. Thus, we should
already start thinking about refactoring xain-fl.coordinator such that
the state machine is completely decoupled for the IO logic.

* XP-308: run isort, black and mypy on the tests package

This is not directly related to task but was done as part of writing
the tests for the storage logic.

Refs:

https://xainag.atlassian.net/browse/XP-40
https://xainag.atlassian.net/browse/XP-308
https://xainag.atlassian.net/browse/XP-309
https://xainag.atlassian.net/browse/XP-311

* XP-308: Add unit test doing an entire round

Refs:

https://xainag.atlassian.net/browse/XP-40
https://xainag.atlassian.net/browse/XP-308
https://xainag.atlassian.net/browse/XP-309
https://xainag.atlassian.net/browse/XP-311

Summary:

This is the first test where multiple participants connect and do a
round of training.

Implementation notes:

- Because we use the peer address to identify a participant, we have to
  use a port forwarding trick to make distinct connections to the
  coordinator. This is implemented in the `tests.port_forwarding` module

- To test the store implementation, we use a custom TestStore class that
  writes data in memory instead of using an actual S3 bucket. That way
  our tests don't depend on an external service. This is implemented in
  the `tests.store` module.

* XP-308: docker: install git and make pip verbose

Git is needed to fetch package from git repositories (xain-proto for
instance).

This commit also makes `pip install` verbose because the command can
take a very long time to execute when gRPC and protobuf are compiled,
and without any output, it may seem like the command hangs.

* XP-308: don't install dev dependencies in docker

The docutils is a transitive dev dependency that is causing problem.

Specifically, when installing dev dependencies in docker,
`docutils==0.16` gets installed, but `boto3` is incompatible with it
and `xain-fl` crashes at runtime:

```
Traceback (most recent call last):
  File "/usr/local/lib/python3.7/site-packages/pkg_resources/__init__.py", line 583, in _build_master
    ws.require(__requires__)
  File "/usr/local/lib/python3.7/site-packages/pkg_resources/__init__.py", line 900, in require
    needed = self.resolve(parse_requirements(requirements))
  File "/usr/local/lib/python3.7/site-packages/pkg_resources/__init__.py", line 791, in resolve
    raise VersionConflict(dist, req).with_context(dependent_req)
pkg_resources.ContextualVersionConflict: (docutils 0.16 (/usr/local/lib/python3.7/site-packages), Requirement.parse('docutils<0.16,>=0.10'), {'botocore'})
```

I don't know why pip pulls this version of docutils though, since
we don't have any dependency that requires such a high version number:

```
$ pipdeptree -r -p docutils
docutils==0.16
  - botocore==1.13.50 [requires: docutils>=0.10,<0.16]
    - boto3==1.10.48 [requires: botocore>=1.13.48,<1.14.0]
      - xain-fl==0.2.0 [requires: boto3==1.10.48]
    - s3transfer==0.2.1 [requires: botocore>=1.12.36,<2.0.0]
      - boto3==1.10.48 [requires: s3transfer>=0.2.0,<0.3.0]
        - xain-fl==0.2.0 [requires: boto3==1.10.48]
  - m2r==0.2.1 [requires: docutils]
  - readme-renderer==24.0 [requires: docutils>=0.13.1]
    - twine==2.0.0 [requires: readme-renderer>=21.0]
  - Sphinx==2.2.1 [requires: docutils>=0.12]
```

* XP-308: print debugging information for spurious failure

* XP-208 join monitor_thread to confirm response to terminate_event

* XP-208 change mock patch string to object called rather than defined

* XP-208 model sum aggregator as a mock for testing

* XP-208 add test integrating P and C

* XP-208 add identity controller as a mock for test

* XP-208 small typo in id-controller docstring

* XP-208 add coordinator test fixture with mocked components

* XP-208 test for top-level function start_participant

* XP-436 fix bug: coordinator not advertising FINISHED state

* XP-436 revert logging level

* XP-208 join monitor_thread to confirm response to terminate_event

* XP-208 change mock patch string to object called rather than defined

* XP-436 suggested documentation fixes from review

* XP-208 mock the store (following rebase)

* XP-208 fix import

* XP-208 test_start_participant: mark as slow, comment on aggregation

Co-authored-by: Corentin Henry <little-dude@users.noreply.github.com>