PB-381: make the log level configurable
Summary:

Add a new `[logging]` section to the configuration file. For now, this
section only provides a `level` key which can be one of the log levels
supported by default by the Python standard library: notset, debug,
info, warning, error, critical.

In the future, this `[logging]` section could be used to provide more
knobs for logging customatization: enabling colors for the terminal
handler, specifying a path for logging into a file, etc.

References:

https://xainag.atlassian.net/browse/PB-381