PB-104: Fix exception in monitor_hearbeats() (#248)
Summary:

This adds a test reproducing the following exception, and provides a
fix for it:

```
Exception in thread Thread-1:

Traceback (most recent call last):
  File "/usr/local/lib/python3.6/threading.py", line 916, in _bootstrap_inner
    self.run()
  File "/usr/local/lib/python3.6/threading.py", line 864, in run
    self._target(*self._args, **self._kwargs)
  File "/app/xain_fl/coordinator/heartbeat.py", line 45, in monitor_heartbeats
    time.sleep(next_expiration)
ValueError: sleep length must be non-negative
```

Detailed explanation:

The is a timing issue in the `monitor_hearbeats()` function, that is
more likely to occur when we have many participants. Here is a
simplified and commented version of the code that explain the issue in
more details:

```python
class Participants:
    def next_expiration(self) -> float:
        return min([p.heartbeat_expires for p in self.participants.values()])

def monitor_hearbeats():
    while True:

        # Here we can have a participant P such that:
        # P.heartbeat_expires > time.time(), but by a very short margin.
        #
        # Thus, P is not removed from the set of participants
        #
        for participant in coordinator.participants.participants.values():
            if participant.heartbeat_expires < time.time():
                coordinator.remove_participant(participant.participant_id)

        # When we arrive here, P is the next participant to expire, so
        # Participants.next_expiration() returns P.heartbeat.expires.
        # Except now, we have P.heartbeat_expires < time.time(). Note
        # that this is more likely to occur if P is at the beginning
        # of the self.coordinator.participants list, and there are
        # many participants, because then, the loop above takes a long
        # time to execute.
        #
        # Thus, next_expiration is negative
        #
        next_expiration: float = coordinator.participants.next_expiration() - time.time()

        # This now fails
        time.sleep(next_expiration)
```

One solution is to only call `time.time()` once in the `while` loop
body. Then, all the durations will be relative to a fixed offset.

Another solution would be simply not to sleep if `next_iterations` is
negative but the fix mentioned ealier seems cleaner.