PB-104: make the tests fast again (#252)
* PB-104: make the tests fast again

Summary:

In https://github.com/xainag/xain-fl/pull/248, we replaced the call to
`time.sleep()` in `monitor_hearbeats()` by a call to
`threading.Event.wait()` to optimize for cases where the termination
event is set while we're sleeping. But, that made tests with mocks of
`time.sleep()` slower. This commit fixes that by mocking
`threading.Event.wait()` instead of `time.sleep()` in these tests.

References:

https://github.com/xainag/xain-fl/pull/248
https://xainag.atlassian.net/browse/PB-104