PB-116: pin docutils version
Summary:

It seems that pip resolves the dependencies by profile. If the [doc]
profile is installed first, docutils ~= 0.16 is picked up, and when
xain-fl's install_requires is resolved, pip sees that docutils is
already installed but has an incompatible version.

I suspect the order in which pip handles the profiles depends on your
environment (pip version, platform, python version, etc.) which is why
we don't see this issue consistently.

As a workaround we explicitely depends on docutils in the [doc]
profile and restrict the version here.

References:

https://xainag.atlassian.net/browse/PB-116