aggregator: when a task finishes cancel the other ones
The aggregator is made of several tasks:
    - the actual Aggregator that run in a background thread
    - the AggregatorService, which reacts to incoming requests from
      participants and RPC requests from the coordinator
    - the API task that handles HTTP requests from participants and
      forwards them to the AggregatorService

If one of these tasks terminates we want to cancel the other ones
because they are useless by themself.

This commit uses the `tokio::select!` macro to wait for one of these
tasks to finish and cancel the other ones. Note that the Aggregator
task itself if reduces to awaiting a Receiver. Cancelling that task
does not terminate the thread in which the Aggregator is running. But
that thread is automatically terminated when the main thread
exits. I'm not sure whether this is clean enough. We might need some
logic to shut the Aggregator down gracefully.