docker: use multi-stages dockerfiles
The image we currently produce is over 4GB large. In order to reduce
this, we perform the compilation in a build container, and just copy
the coordinator and aggregator binaries to the final image.

This commit also removes the dependency vendoring. The main reason for
having it was to cache the dependencies. But that doesn't cache the
builds. Instead we use `cargo fetch` and `cargo build` on an empty
project that contains all our dependencies to force cargo to cache
them.

Finally, we split the images:

- an image to upload to docker hub, that contains a release build
- an image for development that contains a debug build as well as some
  dev tools (Valgrind)