coordinator refactoring
Split the `freeze_dict_xxx()` methods and introduce a
`maybe_transition()` method.

`freeze_dict_xxx()` split
-------------------------

These methods are small but crucial, and actually pretty complex. The
complexity partially stems from the fact that they do two things:
they (1) check if the coordinator is ready to transition to the next
phase, and if so (2) update the coordinator state.

`maybe_transition()`
--------------------

There is currently no method that automatically updates the
coordinator state: the user has to check the current phase, call the
apprporiate `proceed_xxx()` method, and handle the
error. `maybe_transition()` is a function that does all this
automatically.

Tests
-----

The tests are updated to match the changes we made but we also removed
some randomnes from the tests to make the results more reproducible.