coordinator: introduce type aliases & some renaming
Type aliases help the reader understanding the nature of the data
being manipulated.

This commit also renames the `Coordinator.dict_seed` attribute to
`Coordinator.dict_update`. There are two types of seed dictionaries:
the dictionaries sent by the update participant, and the dictionary
built by the coordinator from these updates. It is easy to confuse
them, so we chose to name the dictionaries sent by the participants
"seed dictionary" and the global dictionary built by the coordinator
during the update phase "update dictionary".

Finally, the renaming aforementioned lead to renaming
`Coordinator.update_dict_seed` to `Coordinator.update_dict_update`
which sounds weird, so we renamed these methods as well:

- update_sum_dict => add_sum_participant
- update_dict_seed => add_seed_dict
- update_dict_mask => add_mask_hash