minimalistic service implementation
This commit introduce a service module. The main type is
`service::Service`. It is a Future that handles events (incoming
messages, protocol state changes, etc.), and drives the protocol.

Changes to the protocol state machine
-------------------------------------

To update the state machine, the service calls
`Coordinator::handle_message()` when new messages arrive. The
`Coordinator` updates its state accordingly and emits `ProtocolEvent`
events for the service to handle. Currently, the coordinator emits an
event every time it starts a new phase.

Service state
-------------

The service needs to keep track of some data. This is what
`service::Data` is for. Some data is common to all the protocol
phases, but some is specific to the current protocol phase. The pase
specific data is represented by a `PhaseData` enum.

Service `Handle`
----------------

The service will be running on the tokio event loop, so it won't be
accessible directly. To send events to the service, we defined a
`service::Handle` struct that is just a wrapper around tokio channels.