refactor masking
This commit is very large, so I can't list of the changes that were
made. Here are the most important ones.

More compact BigUint encoding for mask objects
----------------------------------------------

Instead encoding each integer with its length, encode all the integers
with on the same number of bytes. This doesn't waste too much space
due to the random nature of data: the probability that we waste the
`n` most significant bits is 1/2^n.

Do not use generics to encode the primitive type used in the models
-------------------------------------------------------------------

Using `Model<N>` forces us to chose at compile time whether our models
will use f32, f64, i32 or i64. Such restriction is not acceptable for
our use case.

**Important**: This PR does not completely get rid of `Model<N>` yet!
It only limits its reach: for instance the coordinator is not generic
over `N` anymore.

Do not use macros to implement getters/setters
----------------------------------------------

Getters/setters are not idiomatic in Rust. See this reddit discussion:
https://www.reddit.com/r/rust/comments/6etrr1/a_derive_for_your_basic_getters_and_setters/

Make cloning of models and mask objects more explicit
-----------------------------------------------------

- Use APIs that take ownership of mask object and models: masking,
  unmasking and aggregation are operations that _consume_ their
  input. If the caller want to copy the data before-hand it is their
  responsability
- Use iterators instead of vectors, for instance for model
  conversion. Currently we still collect these iterators into vectors,
  but in the future we could try pass them around. `Iterator.map`
  being lazy, this could result in decent performance gains.

Using iterators everywhere would still require a lot of work:
currently, mask objects are vectors for instance. This commit is just
a first step in that direction.