purge pending requests at the end of each phase
Instead of adding the same code at the end of each phase, we leveraged
the existing `Phase` trait. The `Phase::next` method is now split in
two parts:

```rust
trait Phase<R> {
    async fn run(&mut self) -> Result<(), StateError>;
    fn next(self) -> Option<StateMachine<R>>;
}
```

The `run` method runs the phase, and the `next` method just consumes
the phase to produce the next one. By separating these two bits of
logic, we can inject code right after the phase ran to completion
and before the transtion is happens:

```rust
impl<R, S> PhaseState<R, S>
    where Self: Phase<R>
{
    pub async fn run_phase(mut self) -> Option<StateMachine<R>> {
           self.run().await.unwrap();
           do_custom_stuff_common_to_all_the_phases();
           self.next()
    }
}
```

In our case, `do_custom_stuff_common_to_all_the_phases()` consists in
purging the request channel. Again, this is implemented for all the
`PhaseState` via the new `Purge` trait, which is very similar to the
`Handler` trait:

```rust
pub trait Purge<R> {
    fn handle_outdated_request(&mut self, req: R);
}
```

Finally, because we needed to be able to know the current phase, we
added a bunch of `is_xxx()` methods to the `Phase` trait, that return
`true` if the state machine is in the specific phase.