automatically send the phase event before the phase starts (#474)
this started as a small improvement: instead of repeating the same
snippet for each phase, we wanted to do it once in
`PhaseState<R,S>::run_phase()`:

```rust
pub async fn run_phase(mut self) -> Option<StateMachine<R>> {
    let phase = <Self as Phase<R>>::NAME;
    self.coordinator_state.events.broadcast_phase(phase); // <-- add this
    self.run().await.unwrap();
    self.purge_outdated_requests().unwrap();
    self.next()
}
```

It turns out that this doesn't work for the idle phase, because the
round ID is updated in the `run()` implementation, _after_ the phase
event is broadcasted. Therefore, the event that is emitted has the
wrong round id.

There are multiple solutions:

  1. re-emit an event in the idle phase. That may work but that means
     there's a small interval of time during which the listeners still
     have an inconsistent Phase event.

  2. add some logic before broadcasting the phase, such that if this
     is the idle phase, update the round seed before anything
     else. But this logic really belong to the phase itself, so that
     feels very hacky

  3. generate the round seed in `PhaseState<R, Idle>::new`. But
     generating a round seed here is again a bit hacky. This should
     happen when the phase _runs_, in the `Handle::run` impl.

Therefore we went for a 4th solution, and introduced a round id that
is only used internally and has nothing to do with the round
parameters. Updating it in `PhaseState<R, Idle>::new` is fine, since
it has nothing to do with the _phase_ logic itself. The round
parameters are still fully taken care of in `Handle::run`.

Interestingly, this has opened the door to some simplifications in the
`EventPublisher` logic, and we were able to make its API simpler by
removing the round id from the methods arguments altogether.