update to the new message API and refactor PreProcessor
As part of updating the code to match the new message API, I realized
the `PreProcessor` was un-necessarily complex and contained a lot of
duplicated code (once for each message type). The service was
initially intended to perform a bunch of checks on the incoming
messages before passing them to the state machine, but it turns out
the only this it actually does is checking the eligibility of the
participant for the task it wants to take part to. As such we:

- removed the duplicated code
- update the service from `PreProcessor` to `TaskValidator`