add a multipart flags in the message header
The multipart nature of the message is currently encoded in the
message tag. Therefore, the "real" type (sum/update/sum2) of the
message isn't known. This causes multiple problems. First we can't
parse a multipart message without this information. Second, we rely on
the tag information to filter out incoming messages (sum messages are
only accepted during the sum phase for instance). Without this
information, we have to process all the multipart messages we receive.

Therefore this commit restores the flag field in the message header
and use it to indicate whether a message is multipart, and keep the
tag field to indicate the type of the message.