implement MessageEncoder
The `MessageEncoder` is an iterator that takes a message payload of
type `Payload::{Sum,Update,Sum2}`, and yields serialized and signed
messages.

Given a maximum payload length, the encoder automatically chunks the
messages that are too big. For payloads that are small enough to fix
in a single message, the iterator only yields one item. For payload
that needs to be split in several messages, the encoder yields several
items.

Thus, it offers a common serialization interface for both multipart
and non-multipart messages.