refactor xaynet-client
The core of the new client API is the xaynet_sdk::StateMachine. The
state machine implements all the PET logic. Users of the SDK customize
their participants by providing the state machine with types that
implement the following traits:

- `XaynetClient`: a client for communicating with the xaynet
  server. We provide a client based on `reqwest` when the library is
  compile with `--features reqwest_client`
- `Notify`: this is used by the state machine for emitting
  notifications upon notable PET events that may be of interest for
  the participant: selection for the sum or update task, and start of
  a new round. If the participant is not interested in any event,
  users of the library can use the `DefaultNotifier`, which just
  discard the notifications.
- `ModelStore`: this trait is crucial, but also kind of a footgun in
  some ways. We need to document it a bit more carefully. The idea is
  that when the participant is selected for the update task, the state
  machine will poll the store, trying to load the model that the
  participant trained. However, it is the user responsability to
  populate the store with the right model. Not populating the store
  would result in update participants not sending a model, and the
  round to fail. Populating the store with outdated models would
  lead to bad global models.

Note that the state machine is written in such a way that it can be
driven forward by small increments. After each state change, the state
machine yields to the user. The user can thus decide on the strategy
for driving the state machine. In particular, the state machine being
serializable, it is easy to save it and restore it later, in case we
want to pause the execution for a long time.